requirements.txt
langchain-core
langchain-community
langchain-text-splitters
langchain-groq
chromadb
streamlit
tiktoken
pypdf
sentence-transformers
python-dotenv